requirements.txt
requests==2.22.0
requests-toolbelt==0.9.1
ebooklib==0.17.1
addict==2.2.1
attrs==19.3.0
click==7.0
colorama==0.4.3
atomicwrites==1.3.0
python-dateutil==2.8.1
asks==2.4.12
trio==0.19.0
rich==10.16.1